Remove run.sh#file_content_to_terraform_list: we process terraform templates directly
This function was created as roundabout way to deal with terraform
templates where we created the template in a temporary file use
this function to generate terraform list. The upside of this
approach is we avoid the template->template-vars dance. the
downside is we have to remember to run this function everytime a
template file changes, but we often forget.